requirements.txt
Package                                  Version
---------------------------------------- ----------------
adafruit-circuitpython-ads1x15           
numpy
matplotlib
collections
gpiozero
RPi-GPIO
